Services/requirements.txt
boto3
botocore
python-dotenv 
pymongo
motor
fastapi
uvicorn
beanie
pydantic
python-multipart
asgi_lifespan
mongomock_motor
python-ulid[pydantic]
pydantic[email]
langchain
langchain-core
langchain-openai
langchain-community
markdown
requests
httpx
python-jose
passlib
bcrypt
cryptography
google-auth
stripe
Babel
langchain-groq
langchain-google-genai 
google-generativeai
numpy
pandas
